+++ geohabnet/R/params.R
@@ -217,8 +217,37 @@ neg_expo <- function(params = load_parameters(), gammas = NULL, mets = NULL, we 
   return(system.file("defaultParams.yaml", package = "geohabnet", mustWork = TRUE))
 }
 
+.identify.OS = function(){
+  pl <- .Platform$OS.type
+  if(tolower(pl) == "windows"){
+    os = structure("windows", class = "character")
+  }
+  else{
+    si <- as.list(Sys.info())
+    if(tolower(si$sysname) == "linux"){
+      os = structure("Linux", class = "character")
+    }
+    if(tolower(si$sysname) == "darwin"){
+      os = structure("MacOSX", class = "character")
+    }
+  }
+  return(os)
+}
+
 .get_directoryfromuser <- function() {
-  return(easycsv::choose_dir())
+  os = .identify.OS()
+  if(tolower(os) == "windows"){
+    directory <- utils::choose.dir()
+  }
+  if(tolower(os) == "linux"){
+    directory <- system("zenity --file-selection --directory", intern = TRUE)
+  }
+  if(tolower(os) == "macosx"){
+    system("osascript -e 'tell app \"RStudio\" to POSIX path of (choose folder with prompt \"Choose Folder:\")' > /tmp/R_folder",
+           intern = FALSE, ignore.stderr = TRUE)
+    directory <- system("cat /tmp/R_folder && rm -f /tmp/R_folder", intern = TRUE)
+  }
+  return(directory)
 }
 
 .open_file_selection_prompt <- function() {