+++ geohabnet/NEWS.md
@@ -4,6 +4,7 @@
 - Moved Vignettes to articles
 - Support for `terra::rast()`
 - Modified input structure in `parameters.yaml`
+- `host_density->()` is now `habitat_density()->`
 
 # geohabnet 2.1.3
 - Network density plot aviailable as ndplot() for non-global analysis